remove LockedTask
we recently changed the permissions of course teachers to be allowed to
access private tasks within their courses. Due to that change, there is
no longer a need for a class representing tasks the teacher may know
about, but not open.
