+++ apps/server/src/modules/learnroom/controller/dto/board/board-lesson.response.ts
@@ -1,0 +1,32 @@
+import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
+import { DecodeHtmlEntities } from '@shared/controller';
+
+export class BoardLessonResponse {
+	constructor({ id, name, hidden, createdAt, updatedAt }: BoardLessonResponse) {
+		this.id = id;
+		this.name = name;
+		this.hidden = hidden;
+		this.createdAt = createdAt;
+		this.updatedAt = updatedAt;
+	}
+
+	@ApiProperty()
+	id: string;
+
+	@ApiProperty()
+	@DecodeHtmlEntities()
+	name: string;
+
+	@ApiPropertyOptional()
+	@DecodeHtmlEntities()
+	courseName?: string;
+
+	@ApiProperty()
+	createdAt: Date;
+
+	@ApiProperty()
+	updatedAt: Date;
+
+	@ApiProperty()
+	hidden: boolean;
+}
