+++ apps/server/src/modules/learnroom/controller/dto/board/board.response.ts
@@ -1,0 +1,34 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { BoardElementResponse } from './board-element.response';
+
+// TODO: this and DashboardResponse should be combined
+export class BoardResponse {
+	constructor({ roomId, title, displayColor, elements }: BoardResponse) {
+		this.roomId = roomId;
+		this.title = title;
+		this.displayColor = displayColor;
+		this.elements = elements;
+	}
+
+	@ApiProperty({
+		description: 'The id of the room this board belongs to',
+		pattern: '[a-f0-9]{24}',
+	})
+	roomId: string;
+
+	@ApiProperty({
+		description: 'Title of the Board',
+	})
+	title: string;
+
+	@ApiProperty({
+		description: 'Color of the Board',
+	})
+	displayColor: string;
+
+	@ApiProperty({
+		type: [BoardElementResponse],
+		description: 'Array of board specific tasks or lessons with matching type property',
+	})
+	elements: BoardElementResponse[];
+}
