+++ apps/server/src/modules/learnroom/controller/dto/board/index.ts
@@ -1,0 +1,5 @@
+export * from './board-element.response';
+export * from './board-lesson.response';
+export * from './board-task-status.response';
+export * from './board-task.response';
+export * from './board.response';
