+++ apps/server/src/modules/learnroom/controller/dto/index.ts
@@ -1,8 +1,7 @@
+export * from './board';
 export * from './course-metadata.response';
 export * from './dashboard.response';
 export * from './move-position.params';
 export * from './patch-group.params';
 export * from './patch-order.params';
-export * from './board-task.response';
-export * from './board-task-status.response';
 export * from './patch-visibility.params';
