+++ apps/server/src/modules/learnroom/controller/rooms.controller.spec.ts
@@ -1,9 +1,9 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { EntityId, ICurrentUser } from '@shared/domain';
 import { RoomBoardResponseMapper } from '../mapper/room-board-response.mapper';
-import { RoomsUc } from '../uc/rooms.uc';
 import { RoomBoardDTO } from '../types';
-import { BoardResponse } from './dto/board-task.response';
+import { RoomsUc } from '../uc/rooms.uc';
+import { BoardResponse } from './dto';
 import { RoomsController } from './rooms.controller';
 
 describe('rooms controller', () => {
