+++ apps/server/src/modules/learnroom/types/room-board.types.ts
@@ -1,4 +1,4 @@
-import { TaskWithStatusVo, Lesson, EntityId } from '@shared/domain';
+import { Lesson, TaskWithStatusVo } from '@shared/domain';
 
 export type RoomBoardDTO = {
 	roomId: string;
@@ -7,18 +7,12 @@ export type RoomBoardDTO = {
 	elements: RoomBoardElementDTO[];
 };
 
-export type LockedTaskDTO = {
-	id: EntityId;
-	name: string;
-};
-
 export enum RoomBoardElementTypes {
 	TASK = 'task',
-	LOCKEDTASK = 'lockedtask',
 	LESSON = 'lesson',
 }
 
 export type RoomBoardElementDTO = {
 	type: RoomBoardElementTypes;
-	content: TaskWithStatusVo | LockedTaskDTO | Lesson;
+	content: TaskWithStatusVo | Lesson;
 };
