+++ apps/server/src/modules/learnroom/uc/rooms.authorisation.service.spec.ts
@@ -1,7 +1,7 @@
 import { MikroORM } from '@mikro-orm/core';
 import { NotImplementedException } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
-import { taskFactory, courseFactory, lessonFactory, userFactory, setupEntities } from '@shared/testing';
+import { courseFactory, lessonFactory, setupEntities, taskFactory, userFactory } from '@shared/testing';
 import { RoomsAuthorisationService } from './rooms.authorisation.service';
 
 describe('rooms authorisation service', () => {
@@ -91,159 +91,6 @@ describe('rooms authorisation service', () => {
 		});
 	});
 
-	describe('hasTaskListPermission', () => {
-		describe('when task has no course', () => {
-			it('should be true for creator', () => {
-				const user = userFactory.buildWithId();
-				const task = taskFactory.buildWithId({ creator: user, course: undefined });
-
-				const result = service.hasTaskListPermission(user, task);
-				expect(result).toEqual(true);
-			});
-
-			it('should be false for other users', () => {
-				const user = userFactory.buildWithId();
-				const task = taskFactory.buildWithId({ course: undefined });
-
-				const result = service.hasTaskListPermission(user, task);
-				expect(result).toEqual(false);
-			});
-		});
-
-		describe('when task belongs to course', () => {
-			it('should be false for user not in course', () => {
-				const user = userFactory.buildWithId();
-				const course = courseFactory.buildWithId();
-				const task = taskFactory.buildWithId({ course });
-
-				const result = service.hasTaskListPermission(user, task);
-				expect(result).toEqual(false);
-			});
-
-			describe('when task is private', () => {
-				it('should be true for creator', () => {
-					const user = userFactory.buildWithId();
-					const course = courseFactory.buildWithId({ teachers: [user] });
-					const task = taskFactory.draft().buildWithId({ creator: user, course });
-
-					const result = service.hasTaskListPermission(user, task);
-					expect(result).toEqual(true);
-				});
-
-				it('should be true for other teacher', () => {
-					const user = userFactory.buildWithId();
-					const course = courseFactory.buildWithId({ teachers: [user] });
-					const task = taskFactory.draft().buildWithId({ course });
-
-					const result = service.hasTaskListPermission(user, task);
-					expect(result).toEqual(true);
-				});
-
-				it('should be true for other substitutionTeacher', () => {
-					const user = userFactory.buildWithId();
-					const course = courseFactory.buildWithId({ substitutionTeachers: [user] });
-					const task = taskFactory.draft().buildWithId({ course });
-
-					const result = service.hasTaskListPermission(user, task);
-					expect(result).toEqual(true);
-				});
-
-				it('should be false for other student', () => {
-					const user = userFactory.buildWithId();
-					const course = courseFactory.buildWithId({ students: [user] });
-					const task = taskFactory.draft().buildWithId({ course });
-
-					const result = service.hasTaskListPermission(user, task);
-					expect(result).toEqual(false);
-				});
-			});
-
-			describe('when task is unpublished', () => {
-				it('should be true for creator', () => {
-					const user = userFactory.buildWithId();
-					const course = courseFactory.buildWithId({ teachers: [user] });
-					const futureDate = Date.now() + 10000;
-					const task = taskFactory.buildWithId({ creator: user, course, availableDate: futureDate });
-
-					const result = service.hasTaskListPermission(user, task);
-					expect(result).toEqual(true);
-				});
-
-				it('should be true for other teacher', () => {
-					const user = userFactory.buildWithId();
-					const course = courseFactory.buildWithId({ teachers: [user] });
-					const futureDate = Date.now() + 10000;
-					const task = taskFactory.buildWithId({ course, availableDate: futureDate });
-
-					const result = service.hasTaskListPermission(user, task);
-					expect(result).toEqual(true);
-				});
-
-				it('should be true for other substitutionTeacher', () => {
-					const user = userFactory.buildWithId();
-					const course = courseFactory.buildWithId({ substitutionTeachers: [user] });
-					const futureDate = Date.now() + 10000;
-					const task = taskFactory.buildWithId({ course, availableDate: futureDate });
-
-					const result = service.hasTaskListPermission(user, task);
-					expect(result).toEqual(true);
-				});
-
-				it('should be false for other student', () => {
-					const user = userFactory.buildWithId();
-					const course = courseFactory.buildWithId({ students: [user] });
-					const futureDate = Date.now() + 10000;
-					const task = taskFactory.buildWithId({ course, availableDate: futureDate });
-
-					const result = service.hasTaskListPermission(user, task);
-					expect(result).toEqual(false);
-				});
-			});
-
-			describe('when task is published', () => {
-				it('should be true for creator', () => {
-					const user = userFactory.buildWithId();
-					const course = courseFactory.buildWithId({ teachers: [user] });
-					const futureDate = Date.now() - 10000;
-					const task = taskFactory.buildWithId({ creator: user, course, availableDate: futureDate });
-
-					const result = service.hasTaskListPermission(user, task);
-					expect(result).toEqual(true);
-				});
-
-				it('should be true for other teacher', () => {
-					const user = userFactory.buildWithId();
-					const course = courseFactory.buildWithId({ teachers: [user] });
-					const futureDate = Date.now() - 10000;
-					const task = taskFactory.buildWithId({ course, availableDate: futureDate });
-
-					const result = service.hasTaskListPermission(user, task);
-					expect(result).toEqual(true);
-				});
-
-				it('should be true for other substitutionTeacher', () => {
-					const user = userFactory.buildWithId();
-					const course = courseFactory.buildWithId({ substitutionTeachers: [user] });
-					const futureDate = Date.now() - 10000;
-					const task = taskFactory.buildWithId({ course, availableDate: futureDate });
-
-					const result = service.hasTaskListPermission(user, task);
-					expect(result).toEqual(true);
-				});
-
-				it('should be true for other student', () => {
-					const user = userFactory.buildWithId();
-					const course = courseFactory.buildWithId({ students: [user] });
-					const pastDate = Date.now() - 10000;
-					const task = taskFactory.buildWithId({ course, availableDate: pastDate });
-
-					const result = service.hasTaskListPermission(user, task);
-					expect(result).toEqual(true);
-				});
-			});
-		});
-	});
-
 	describe('hasTaskReadPermission', () => {
 		describe('when task has no course', () => {
 			it('should be true for creator', () => {
@@ -283,22 +130,22 @@ describe('rooms authorisation service', () => {
 					expect(result).toEqual(true);
 				});
 
-				it('should be false for other teacher', () => {
+				it('should be true for other teacher', () => {
 					const user = userFactory.buildWithId();
 					const course = courseFactory.buildWithId({ teachers: [user] });
 					const task = taskFactory.draft().buildWithId({ course });
 
 					const result = service.hasTaskReadPermission(user, task);
-					expect(result).toEqual(false);
+					expect(result).toEqual(true);
 				});
 
-				it('should be false for other substitutionTeacher', () => {
+				it('should be true for other substitutionTeacher', () => {
 					const user = userFactory.buildWithId();
 					const course = courseFactory.buildWithId({ substitutionTeachers: [user] });
 					const task = taskFactory.draft().buildWithId({ course });
 
 					const result = service.hasTaskReadPermission(user, task);
-					expect(result).toEqual(false);
+					expect(result).toEqual(true);
 				});
 
 				it('should be false for other student', () => {
@@ -447,14 +294,14 @@ describe('rooms authorisation service', () => {
 		});
 	});
 
-	describe('hasLessonListPermission', () => {
+	describe('hasLessonReadPermission', () => {
 		describe('when lesson is hidden', () => {
 			it('should be true for teacher of course', () => {
 				const user = userFactory.buildWithId();
 				const course = courseFactory.buildWithId({ teachers: [user] });
 				const lesson = lessonFactory.buildWithId({ course, hidden: true });
 
-				const result = service.hasLessonListPermission(user, lesson);
+				const result = service.hasLessonReadPermission(user, lesson);
 				expect(result).toEqual(true);
 			});
 
@@ -463,7 +310,7 @@ describe('rooms authorisation service', () => {
 				const course = courseFactory.buildWithId({ substitutionTeachers: [user] });
 				const lesson = lessonFactory.buildWithId({ course, hidden: true });
 
-				const result = service.hasLessonListPermission(user, lesson);
+				const result = service.hasLessonReadPermission(user, lesson);
 				expect(result).toEqual(true);
 			});
 
@@ -472,7 +319,7 @@ describe('rooms authorisation service', () => {
 				const course = courseFactory.buildWithId({ students: [user] });
 				const lesson = lessonFactory.buildWithId({ course, hidden: true });
 
-				const result = service.hasLessonListPermission(user, lesson);
+				const result = service.hasLessonReadPermission(user, lesson);
 				expect(result).toEqual(false);
 			});
 
@@ -481,7 +328,7 @@ describe('rooms authorisation service', () => {
 				const course = courseFactory.buildWithId();
 				const lesson = lessonFactory.buildWithId({ course, hidden: true });
 
-				const result = service.hasLessonListPermission(user, lesson);
+				const result = service.hasLessonReadPermission(user, lesson);
 				expect(result).toEqual(false);
 			});
 		});
@@ -492,7 +339,7 @@ describe('rooms authorisation service', () => {
 				const course = courseFactory.buildWithId({ teachers: [user] });
 				const lesson = lessonFactory.buildWithId({ course });
 
-				const result = service.hasLessonListPermission(user, lesson);
+				const result = service.hasLessonReadPermission(user, lesson);
 				expect(result).toEqual(true);
 			});
 
@@ -501,7 +348,7 @@ describe('rooms authorisation service', () => {
 				const course = courseFactory.buildWithId({ substitutionTeachers: [user] });
 				const lesson = lessonFactory.buildWithId({ course });
 
-				const result = service.hasLessonListPermission(user, lesson);
+				const result = service.hasLessonReadPermission(user, lesson);
 				expect(result).toEqual(true);
 			});
 
@@ -510,7 +357,7 @@ describe('rooms authorisation service', () => {
 				const course = courseFactory.buildWithId({ students: [user] });
 				const lesson = lessonFactory.buildWithId({ course });
 
-				const result = service.hasLessonListPermission(user, lesson);
+				const result = service.hasLessonReadPermission(user, lesson);
 				expect(result).toEqual(true);
 			});
 
@@ -519,7 +366,7 @@ describe('rooms authorisation service', () => {
 				const course = courseFactory.buildWithId();
 				const lesson = lessonFactory.buildWithId({ course });
 
-				const result = service.hasLessonListPermission(user, lesson);
+				const result = service.hasLessonReadPermission(user, lesson);
 				expect(result).toEqual(false);
 			});
 		});
