+++ apps/server/src/modules/learnroom/uc/rooms.authorisation.service.ts
@@ -21,7 +21,7 @@ export class RoomsAuthorisationService {
 		return hasPermission;
 	}
 
-	hasTaskListPermission(user: User, task: Task): boolean {
+	hasTaskReadPermission(user: User, task: Task): boolean {
 		const isCreator = task.creator === user;
 		let hasCoursePermission = false;
 
@@ -42,20 +42,7 @@ export class RoomsAuthorisationService {
 		return hasPermission;
 	}
 
-	hasTaskReadPermission(user: User, task: Task): boolean {
-		let hasPermission = false;
-
-		if (task.private) {
-			const isCreator = task.creator === user;
-			hasPermission = isCreator;
-		} else {
-			hasPermission = this.hasTaskListPermission(user, task);
-		}
-
-		return hasPermission;
-	}
-
-	hasLessonListPermission(user: User, lesson: Lesson): boolean {
+	hasLessonReadPermission(user: User, lesson: Lesson): boolean {
 		let hasCoursePermission = false;
 		hasCoursePermission = this.hasCourseReadPermission(user, lesson.course);
 		if (lesson.hidden) {
